Move files under vs/code/electron-browser/issue
