+++ src/vs/code/buildfile.js
@@ -21,6 +21,7 @@ exports.collectModules= function() {
 		createModuleDescription('vs/code/electron-main/main', []),
 		createModuleDescription('vs/code/node/cli', []),
 		createModuleDescription('vs/code/node/cliProcessMain', ['vs/code/node/cli']),
-		createModuleDescription('vs/code/electron-browser/sharedProcess/sharedProcessMain', [])
+		createModuleDescription('vs/code/electron-browser/sharedProcess/sharedProcessMain', []),
+		createModuleDescription('vs/code/electron-browser/issue/issueReporterMain', [])
 	];
 };
