+++ src/vs/code/electron-browser/issue/bootstrap.css
