+++ src/vs/code/electron-browser/issue/issueReporter.css
