+++ src/vs/code/electron-browser/issue/issueReporter.html
@@ -99,5 +99,5 @@
 		</div>
 	</body>
 
-	<script src="index.js"></script>
+	<script src="issueReporter.js"></script>
 </html>
