+++ src/vs/code/electron-browser/issue/issueReporterModel.ts
